ui/gui fixes

+++ libs/gui/bufferqueue/1.0/H2BGraphicBufferProducer.cpp
@@ -127,7 +127,7 @@ inline void wrapAs(AnwBuffer* t, GraphicBuffer const& l) {
     t->attr.stride = l.getStride();
     t->attr.format = static_cast<PixelFormat>(l.getPixelFormat());
     t->attr.layerCount = l.getLayerCount();
-    t->attr.usage = l.getUsage();
+    t->attr.usage = uint32_t(l.getUsage());     // FIXME: need 64-bits usage version
     t->attr.id = l.getId();
     t->attr.generationNumber = l.getGenerationNumber();
     t->nativeHandle = hidl_handle(l.handle);
@@ -993,12 +993,12 @@ status_t H2BGraphicBufferProducer::setAsyncMode(bool async) {
 // FIXME: usage bits truncated -- needs a 64-bits usage version
 status_t H2BGraphicBufferProducer::dequeueBuffer(int* slot, sp<Fence>* fence, uint32_t w,
                                                  uint32_t h, ::android::PixelFormat format,
-                                                 uint32_t usage, uint64_t* outBufferAge,
+                                                 uint64_t usage, uint64_t* outBufferAge,
                                                  FrameEventHistoryDelta* outTimestamps) {
     *fence = new Fence();
     status_t fnStatus;
     status_t transStatus = toStatusT(mBase->dequeueBuffer(
-            w, h, static_cast<PixelFormat>(format), usage,
+            w, h, static_cast<PixelFormat>(format), uint32_t(usage),
             outTimestamps != nullptr,
             [&fnStatus, slot, fence, outTimestamps] (
                     Status status,
@@ -1151,10 +1151,11 @@ status_t H2BGraphicBufferProducer::setSidebandStream(
     return toStatusT(mBase->setSidebandStream(stream == nullptr ? nullptr : stream->handle()));
 }
 
+// FIXME: usage bits truncated -- needs a 64-bits usage version
 void H2BGraphicBufferProducer::allocateBuffers(uint32_t width, uint32_t height,
-        ::android::PixelFormat format, uint32_t usage) {
+        ::android::PixelFormat format, uint64_t usage) {
     mBase->allocateBuffers(
-            width, height, static_cast<PixelFormat>(format), usage);
+            width, height, static_cast<PixelFormat>(format), uint32_t(usage));
 }
 
 status_t H2BGraphicBufferProducer::allowAllocation(bool allow) {
@@ -1233,14 +1234,14 @@ status_t H2BGraphicBufferProducer::getUniqueId(uint64_t* outId) const {
     return transStatus == NO_ERROR ? fnStatus : transStatus;
 }
 
-status_t H2BGraphicBufferProducer::getConsumerUsage(uint32_t* outUsage) const {
+status_t H2BGraphicBufferProducer::getConsumerUsage(uint64_t* outUsage) const {
     ALOGW("getConsumerUsage is not fully supported");
     int result;
     status_t transStatus = toStatusT(mBase->query(
             NATIVE_WINDOW_CONSUMER_USAGE_BITS,
             [&result, outUsage] (int32_t tResult, int32_t tValue) {
                 result = static_cast<int>(tResult);
-                *outUsage = static_cast<uint32_t>(tValue);
+                *outUsage = static_cast<uint64_t>(tValue);
             }));
     return transStatus == NO_ERROR ? result : static_cast<int>(transStatus);
 }
